sync allows deletion of directories

+++ sync.php
@@ -170,15 +170,22 @@ if(!empty($files)){
 	}
 
 	$dir_iterator = new RecursiveDirectoryIterator("content/remote/");
-	$dir_contents = new RecursiveIteratorIterator($dir_iterator, RecursiveIteratorIterator::SELF_FIRST);
+	$dir_contents = new RecursiveIteratorIterator($dir_iterator, RecursiveIteratorIterator::CHILD_FIRST);
 
-	foreach ($dir_contents as $filename) {
+	foreach ($dir_contents as $filename){
 		$localfile = str_replace("content/remote/", "", $filename);
 		if(!in_array($localfile, $filecheck)){
-			if($localfile != ".dropbox_sync_hash"){
-				//echo "- ".$localfile."<br/>";
+			if($localfile != ".dropbox_sync_hash" &&
+			   $localfile != "." &&
+			   $localfile != ".."){
+				
 				$output = "-";
-				unlink("content/remote/".$localfile);
+
+				if(is_dir("content/remote/".$localfile)){
+					rmdir_recursive("content/remote/".$localfile);
+				} else {
+					unlink("content/remote/".$localfile);
+				}
 			}
 		}
 	}
@@ -187,6 +194,17 @@ if(!empty($files)){
 
 }
 echo $output;
+
+# - - - - - - - - - - - - -
+#
+function rmdir_recursive($dir) {
+    foreach(scandir($dir) as $file) {
+        if ('.' === $file || '..' === $file) continue;
+        if (is_dir("$dir/$file")) rmdir_recursive("$dir/$file");
+        else unlink("$dir/$file");
+    }
+    rmdir($dir);
+}
 
 # - - - - - - - - - - - - -
 #
